requirements.txt
requests
python-dotenv
re
json
langchain_huggingface